Extract component from widget, resurrect smaller compass reading
Create separate files for the actual React component to allow
different ways to render a single logical value. Resurrect older and
simpler and more compact version of the compass as a proof of concept.
